+++ packages/spacecat-shared-utils/CHANGELOG.md
@@ -1,3 +1,17 @@
+# [@adobe/spacecat-shared-utils-v1.14.5](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.14.4...@adobe/spacecat-shared-utils-v1.14.5) (2024-04-16)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#202](https://github.com/adobe/spacecat-shared/issues/202)) ([d5d20a8](https://github.com/adobe/spacecat-shared/commit/d5d20a8d606c7589fb794d6c6e3319541a3e6439))
+
+# [@adobe/spacecat-shared-utils-v1.14.4](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.14.3...@adobe/spacecat-shared-utils-v1.14.4) (2024-04-16)
+
+
+### Bug Fixes
+
+* strip slash only for / path ([#209](https://github.com/adobe/spacecat-shared/issues/209)) ([4c009ea](https://github.com/adobe/spacecat-shared/commit/4c009ea35e85ee160ee025f53aaa3918f32aa0b1))
+
 # [@adobe/spacecat-shared-utils-v1.14.3](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.14.2...@adobe/spacecat-shared-utils-v1.14.3) (2024-04-10)